Fix three pre-existing clippy warnings
- ipc.rs: drop a u64 -> u64 no-op cast (WindowId is a plain u64 alias).
- decoration.rs: draw_line took 8 args; bundle the two endpoints into
  (i32, i32) tuples instead of four loose coordinates.
- output_management.rs: collapse a WEnum::Value if-let directly into
  the outer SetTransform match arm's pattern.

+++ crates/platform/src/ipc.rs
@@ -279,7 +279,7 @@ fn client_snapshot(wm: &std::rc::Rc<std::cell::RefCell<WindowManager>>) -> Vec<C
     let focused = wm.focused_id();
     wm.windows()
         .map(|w| ClientInfo {
-            id: w.id as u64,
+            id: w.id,
             app_id: w.app_id.clone(),
             title: w.title.clone(),
             workspace: w.workspace,
